refactor: DTO 통합 및 자전거 프로필 타입 안정성 개선
- full-journey.dto와 round-trip.dto를 round.dto로 통합
- BikeProfile enum으로 타입 안정성 확보 및 프로필 변환 로직 추가
